Fixed warnings.

+++ src/ios/Utils/NSData+HCPMD5.m
@@ -14,7 +14,7 @@
     unsigned char md5Buffer[CC_MD5_DIGEST_LENGTH];
     
     // Create 16 byte MD5 hash value, store in buffer
-    CC_MD5(self.bytes, self.length, md5Buffer);
+    CC_MD5(self.bytes, (unsigned int)self.length, md5Buffer);
     
     // Convert unsigned char buffer to NSString of hex values
     NSMutableString *output = [NSMutableString stringWithCapacity:CC_MD5_DIGEST_LENGTH * 2];
